+++ src/main/java/com/sqlcsv/sqlcsv/todo/ITableService.java
@@ -1,13 +1,13 @@
-package com.sqlcsv.sqlcsv.service;
+package com.sqlcsv.sqlcsv.todo;
 
 import com.sqlcsv.sqlcsv.model.Table;
 
 import java.io.IOException;
 import java.util.List;
 
-public interface TableServiceInterface {
+public interface ITableService {
     List<String> getAllSpreadsheetsIds() throws IOException;
-    List<String> gettAllSheetsNames(int spreadsheetId);
+    List<String> getAllSheetsNames(int spreadsheetId);
     Table getTableById(int sheetId);
     Table queryTheTable(String query);
 }
