+++ src/main/java/com/sqlcsv/sqlcsv/todo/TableRepositoryInterface.java
@@ -1,4 +1,4 @@
-package com.sqlcsv.sqlcsv.repository;
+package com.sqlcsv.sqlcsv.todo;
 
 import com.sqlcsv.sqlcsv.model.Table;
