+++ src/main/java/com/sqlcsv/sqlcsv/todo/TableRepositoryPredicates.java
@@ -1,4 +1,4 @@
-package com.sqlcsv.sqlcsv.service;
+package com.sqlcsv.sqlcsv.todo;
 
 import com.sqlcsv.sqlcsv.model.Table;
 import org.springframework.stereotype.Component;
